Trait otopr::decoding::Decodable

pub trait Decodable<'de>: Sized {
    type Wire: WireType;
    fn decode<B: Buf>(deserializer: &mut Deserializer<'de, B>) -> Result<Self>;

    fn merge_from<B: Buf>(
        &mut self, 
        deserializer: &mut Deserializer<'de, B>
    ) -> Result<()> { ... }

    fn merge(&mut self, other: Self) { ... }
}

Associated Types

type Wire: WireType

Required methods

fn decode<B: Buf>(deserializer: &mut Deserializer<'de, B>) -> Result<Self>

Provided methods

fn merge_from<B: Buf>(
&mut self,
deserializer: &mut Deserializer<'de, B>
) -> Result<()>

fn merge(&mut self, other: Self)

If this is a message, call merge() on all fields, if this is repeated, extend this with the elements of other. for all other types simply overwrite this with other, which is the default.