Trait otopr::decoding::DecodableMessage

pub trait DecodableMessage<'de>: Sized {
    type Tag: VarInt;
    fn decode_field<B: Buf>(
        &mut self, 
        deserializer: &mut Deserializer<'de, B>, 
        tag: Self::Tag
    ) -> Result<()>;

    fn decode<B: Buf>(deserializer: &mut Deserializer<'de, B>) -> Result<Self>
    where
        Self: Default,
    { ... }
}

Associated Types

type Tag: VarInt

How big the tag message gets. This is an unsigned varint.

It is not an error if any field tag overflows this type, since there can be removed fields exceeding the current storage type.

Required methods

fn decode_field<B: Buf>(
&mut self,
deserializer: &mut Deserializer<'de, B>,
tag: Self::Tag
) -> Result<()>

Decodes a field with the given tag.

Skips the field if there are no matches for the tag.

Provided methods

fn decode<B: Buf>(deserializer: &mut Deserializer<'de, B>) -> Result<Self> where
Self: Default,